refactor(text-input.tsx): refactoring all inputs in small components
