+++ hosting/src/components/custom/textarea-input.tsx
@@ -1,0 +1,19 @@
+import { Field } from 'formik'
+import { Textarea } from '../ui/textarea'
+
+interface InputProps {
+  label: string
+  placeholder?: string
+  id: string
+  name: string
+  titleTextArea?: string
+}
+
+export function TextareaInput({ label, titleTextArea, placeholder, id, name }: InputProps) {
+  return (
+    <div>
+      <label htmlFor={label}>{titleTextArea}</label>
+      <Field placeholder={placeholder} id={id} name={name} as={Textarea} className="resize-none" />
+    </div>
+  )
+}
