+++ hosting/src/routes/password-changes.tsx
@@ -1,4 +1,4 @@
-import { InputForm } from '@/components/custom/generics-inputs'
+import { InputForm } from '@/components/custom/text-input'
 import { changePasswordRequestSchema } from '@/models/change-password-request-schema'
 import { useChangePasswordRequestQuery } from '@/queries/use-change-password-request-query'
 import { useStudentsListQuery } from '@/queries/use-students-list-query'
@@ -49,7 +49,6 @@ export function ChangedPasswordRequests() {
             name="newPasswordDefault"
             placeholder="Digite a senha padrão aqui"
             label="newPasswordDefault"
-            type="text"
           />
 
           <InputForm
@@ -58,7 +57,6 @@ export function ChangedPasswordRequests() {
             name="requestStatus"
             placeholder="Digite o status da requisição de troca de senha"
             label="requestStatus"
-            type="text"
           />
         </Form>
       </Formik>
